refactor(core): Push environment and inputStream to JobRequest
We currently have a JobRequest object that holds a request for a
job to be executed. It currently only contains the command to be
executed, requiring the environment and any input stream to be
separately passed to the execute command. Allow the JobRequest to
specify the environment and inputStream so that it can encapsulate
these parameters as well.  This also makes it easier to default
these parameters, as almost all calls pass system.getenv() as the
environment and an empty stream as the input stream.
