+++ clouddriver-appengine/src/main/groovy/com/netflix/spinnaker/clouddriver/appengine/AppengineJobExecutor.groovy
@@ -32,9 +32,7 @@ class AppengineJobExecutor {
   JobExecutor jobExecutor
 
   void runCommand(List<String> command) {
-    JobStatus jobStatus = jobExecutor.runJob(new JobRequest(command),
-                                        System.getenv(),
-                                        new ByteArrayInputStream())
+    JobStatus jobStatus = jobExecutor.runJob(new JobRequest(command))
     if (jobStatus.getResult() == JobStatus.Result.FAILURE && jobStatus.getStdOut()) {
       String stdOut = jobStatus.getStdOut()
       String stdErr = jobStatus.getStdErr()
